Move page number tests into its own test file
