+++ test/util/string-functions.spec.js
@@ -1,7 +1,6 @@
 const { expect } = require('chai')
 
 const { hasUpperCaseCharacterInMiddleOfWord, normalizedCharCodeArray, removeLeadingWhitespaces, removeTrailingWhitespaces, prefixAfterWhitespace, suffixBeforeWhitespace, charCodeArray, isListItem, isNumberedListItem, wordMatch } = require('../../lib/util/string-functions')
-const { findPageNumbers, findFirstPage } = require('../../lib/util/pdf')
 
 describe('functions: hasUpperCaseCharacterInMiddleOfWord', () => {
   it('single word', () => {
@@ -177,46 +176,5 @@ describe('functions: wordsMatch', () => {
     expect(wordMatch('text', 'test')).to.equal(0.0)
 
     expect(wordMatch('inStruCtionS for the full Moon proCeSS', 'Instructions for the Full Moon Process')).to.equal(1.0)
-  })
-})
-
-describe('functions: findPageNumbers', () => {
-  it('Search, coerce and store page number', () => {
-    const array = [{ str: '1' }, { str: 'test' }, { str: 'how' }, { str: 'to' }, { str: 'find' }, { str: 'page' }, { str: '3' }, { str: 'number' }]
-    expect(findPageNumbers({}, 3, array)).to.eql({ 3: [1, 3] })
-  })
-})
-
-describe.only('functions: findFirstPage', () => {
-  it('Return first page index and number with completed Hashtable', () => {
-    const hashTable = {
-      20: [3],
-      21: [4],
-      22: [5],
-      23: [6],
-      24: [7],
-      25: [8],
-      26: [9],
-      27: [10],
-      28: [11],
-      30: [13],
-    }
-    expect(findFirstPage(hashTable)).to.eql(['20', [3]])
-  })
-
-  it('Return first page index and number with incompleted Hashtable', () => {
-    const hashTable = {
-      2: [19, 86, 1986, 110],
-      5: [137, 151],
-      9: [1],
-      10: [4],
-      12: [6],
-      13: [8, 7],
-      14: [8],
-      15: [10, 9],
-      16: [11, 12, 10],
-      17: [11],
-    }
-    expect(findFirstPage(hashTable)).to.eql(['10', [4]])
   })
 })
